made LowPtGsfElectronIDProducer a global module
